add support for jsxbin v1.0
Apart from new nodes there were only very few changes in v2.0 compared
to v1.0, namely a new flag in v2.0 for IdNodes and References as well as
support for "for each" in v2.0.
The code has been split up with the code that is different depending on
the version marked as such.
Tests also have been added for v1.0 since they are slightly different in
certain places.